+++ controlpanel/auth/__init__.py
@@ -1,0 +1,1 @@
+from controlpanel.auth.oidc import OIDCSessionValidator, OIDCSubAuthenticationBackend, oauth  # noqa
