+++ controlpanel/auth/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path("login/", views.OIDCLoginView.as_view(), name="login"),
+    path("authenticate/", views.OIDCAuthenticationView.as_view(), name="authenticate"),
+    path("logout/", views.LogoutView.as_view(), name="logout"),
+    path("login-fail/", views.LoginFail.as_view(), name="login-fail"),
+]
