+++ controlpanel/auth/views/__init__.py
@@ -1,0 +1,1 @@
+from .views import LoginFail, LogoutView, OIDCAuthenticationView, OIDCLoginView  # noqa
