--- controlpanel/core/auth/__init__.py
@@ -1,5 +1,0 @@
-from controlpanel.core.auth.oidc import (  # noqa
-    OIDCSessionValidator,
-    OIDCSubAuthenticationBackend,
-    oauth,
-)
