--- controlpanel/interfaces/web/auth/__init__.py
@@ -1,6 +1,0 @@
-from controlpanel.interfaces.web.auth.views import (  # noqa
-    LoginFail,
-    LogoutView,
-    OIDCAuthenticationView,
-    OIDCLoginView,
-)
