--- controlpanel/interfaces/web/data_products/__init__.py
@@ -1,1 +1,0 @@
-from controlpanel.interfaces.web.data_products.views import DataProductsView  # noqa
