--- controlpanel/interfaces/web/data_products/views.py
@@ -1,5 +1,0 @@
-from django.views.generic import TemplateView
-
-
-class DataProductsView(TemplateView):
-    template_name = "data-products.html"
