--- controlpanel/interfaces/web/templates/data-products.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-
-{% extends "base.html" %}
-
-{% block content %}
-<h1 class="govuk-heading-xl"> Data products </h1>
-
-{% endblock %}
